Converted to Onion architecture
