+++ src/Chirp.Core/ChirpDBContext.cs
@@ -1,4 +1,4 @@
-using Chirp.Razor.Shared.Storage;
+using Chirp.Web.Storage;
 using Microsoft.EntityFrameworkCore;
 using System;
