+++ src/Chirp.Web/Pages/Public.cshtml
@@ -1,14 +1,12 @@
-﻿@page "/{author}"
-@model Chirp.Razor.Pages.UserTimelineModel
+﻿@page "/"
+@model Chirp.Web.Pages.PublicModel
 @{
     ViewData["Title"] = "Chirp!";
     Layout = "Shared/_Layout";
-    var routeName = HttpContext.GetRouteValue("author");
-
 }
 
 <div>
-    <h2> @routeName's Timeline </h2>
+    <h2> Public Timeline </h2>
 
     @if (Model.Cheeps.Any())
     {
@@ -35,7 +33,7 @@
     <div class="pagination">
         @for (int i = 1; i <= Math.Ceiling((double) Model.NumOfCheeps / Model.CheepsPerPage); i++)
         {
-            <a href="/@routeName/?page=@i">@i</a>
+            <a href="/?page=@i">@i</a>
         }
     </div>
 </div>
