+++ src/Chirp.Web/Pages/Public.cshtml.cs
@@ -1,27 +1,25 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace Chirp.Razor.Pages;
+namespace Chirp.Web.Pages;
 
-public class UserTimelineModel : PageModel
+public class PublicModel : PageModel
 {
     private readonly ICheepService _service;
-    public List<Cheep> Cheeps { get; set; }
+    public List<Cheep> Cheeps { get; set; } = new();
+
     public int CheepsPerPage;
     public int NumOfCheeps;
 
-
-    public UserTimelineModel(ICheepService service)
+    public PublicModel(ICheepService service)
     {
         _service = service;
     }
 
-    public ActionResult OnGet(string author, [FromQuery] int page = 1)
+    public ActionResult OnGet([FromQuery] int page = 1)
     {
-
-        NumOfCheeps = _service.GetCheepCount(author);
-
-        int maxPage = (NumOfCheeps / _service.CheepsPerPage) + 1;
+        NumOfCheeps = _service.GetCheepCount();
+        int maxPage = NumOfCheeps / _service.CheepsPerPage;
 
         if (page == 0)
         {
@@ -33,7 +31,8 @@ public class UserTimelineModel : PageModel
             return RedirectToPage();
         }
 
-        Cheeps = _service.GetCheeps(page, author);
+        Cheeps = _service.GetCheeps(page);
+
         CheepsPerPage = _service.CheepsPerPage;
         return Page();
     }
