+++ src/Chirp.Web/wwwroot/favicon/favicon.ico
